[FIX] 修复主键加密

See merge request hzero-agile/test-manager-service!103

+++ src/main/java/io/choerodon/test/manager/app/service/TestPlanServcie.java
@@ -57,7 +57,7 @@ public interface TestPlanServcie {
      * saga复制计划
      * @param map
      */
-    void sagaClonePlan(Map<String,Integer> map);
+    void sagaClonePlan(Map<String,Long> map);
 
     /**
      * 改变计划的init状态为fail
